Use decimal for score, better regex

+++ interface/models.py
@@ -84,7 +84,9 @@ class Submission(models.Model):
     total_score = models.DecimalField(max_digits=5,
                                       decimal_places=2,
                                       null=True)
-    score = models.IntegerField(null=True)
+    score = models.DecimalField(max_digits=5,
+                                decimal_places=2,
+                                null=True)
     archive_size = models.IntegerField(null=True)
     vmck_job_id = models.IntegerField(null=True)
